refactor(agent-runner): split tool-call normalization ownership
