+++ config/max-lines-baseline.txt
@@ -379,8 +379,6 @@ src/agents/embedded-agent-runner/run/attempt-spawn-workspace.test-support.ts
 src/agents/embedded-agent-runner/run/attempt.spawn-workspace.context-engine.test.ts
 src/agents/embedded-agent-runner/run/attempt.test.ts
 src/agents/embedded-agent-runner/run/attempt.tool-call-argument-repair.ts
-src/agents/embedded-agent-runner/run/attempt.tool-call-normalization.test.ts
-src/agents/embedded-agent-runner/run/attempt.tool-call-normalization.ts
 src/agents/embedded-agent-runner/run/incomplete-turn.ts
 src/agents/embedded-agent-runner/runs.ts
 src/agents/embedded-agent-runner/thinking.test.ts
